requirements.txt
streamlit>=1.31
transformers>=4.40
datasets>=2.18
torch>=2.0
scikit-learn>=1.3

sentencepiece
tf-keras
newspaper3k
lxml_html_clean
